refactor(onboarding): Extract WelcomeSkipButton component
Create reusable WelcomeSkipButton component that encapsulates skip
onboarding logic with analytics tracking. Use it in both welcome
screens.
